+++ lib/common/constants.dart
@@ -37,7 +37,10 @@ class ZOStrings {
 
 class Constants {
   static const lastWeekOfYear = 52;
-  static const pickupConfirmationTime = 45;
+
+  // ZOB-234 Use shorter confirmation time for personal delivery
+  static const pickupConfirmationTimeDefault = 45;
+  static const pickupConfirmationTimePersonal = 20;
 }
 
 class WidgetStyle {