+++ lib/models/delivery.dart
@@ -1,24 +1,33 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:zachranobed/common/constants.dart';
 
 /*
  * Command to rebuild the delivery.g.dart file:
  * flutter pub run build_runner build --delete-conflicting-outputs
  */
 part 'delivery.freezed.dart';
-part 'delivery.g.dart';
 
 @Freezed()
 class Delivery with _$Delivery {
+  const Delivery._();
   const factory Delivery({
     required String id,
     required String donorId,
     required String recipientId,
     required DeliveryState state,
     required DeliveryType type,
+    required CarrierType carrierType,
   }) = _Delivery;
 
-  factory Delivery.fromJson(Map<String, dynamic> json) =>
-      _$DeliveryFromJson(json);
+  /// Returns the pickup confirmation time based on the carrier type.
+  int getConfirmationTime() {
+    switch (carrierType) {
+      case CarrierType.personal:
+        return Constants.pickupConfirmationTimePersonal;
+      case CarrierType.other:
+        return Constants.pickupConfirmationTimeDefault;
+    }
+  }
 }
 
 enum DeliveryState {
@@ -33,4 +42,9 @@ enum DeliveryState {
 enum DeliveryType {
   foodDelivery,
   boxDelivery,
+}
+
+enum CarrierType {
+  personal,
+  other,
 }