+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This is a changelog for **ZachranObed** application.
 - **ZOB-101** Disable auto-correct for login fields.
 - **ZOB-199** Update notification package to 17.1.2.
 - **ZOB-214** Change Android launcher icon to rounded.
+- **ZOB-237** Change pickup confirmation time for personal delivery.
 
 ### Removed